Add files via upload
Version 2 - THis version is still being update to function correctly. HTe idea behind this will allow the user to set the sides if they want to use differnt sided dice. THis would be useful, for example, when rolling for combat in DnD, first roll for hit and second roll for damage.

RTHis version currently need to have the variables updated in the functions to ebsure that everthing works.